Significantly improve debuggability of things and properly make sure that telem works, but I definitely still have a bug where the teensy crashes after reaching > 2000 CCNO

+++ include/state_field_registry.hpp
@@ -43,7 +43,7 @@ struct StateFields
     sfr_gen.header_gen_all()
     ]]]*/
     float time_t_average_cycle_time_us{};
-    uint64_t mcl_control_cycle_num{};
+    uint32_t mcl_control_cycle_num{};
     MainControl::State mc_state{};
     MainControl::State target_mc_state{};
     GncControl::State gnc_state{};
